moved auth handle to it's own middleware class
